feat: move self-signed cert plugin
