+++ pkgs/standards/swarmauri_certs_self_signed/tests/perf/test_self_signed_certificate_perf.py
@@ -1,15 +1,14 @@
 import pytest
 from cryptography.hazmat.primitives import serialization
 from cryptography.hazmat.primitives.asymmetric import rsa
-from cryptography import x509
 
-from swarmauri_standards import SelfSignedCertificate
+from swarmauri_certs_self_signed import SelfSignedCertificate
 from swarmauri_core.crypto.types import KeyRef, KeyType, KeyUse, ExportPolicy
 
 
 @pytest.mark.test
-@pytest.mark.unit
-def test_issue_basic():
+@pytest.mark.perf
+def test_issue_perf(benchmark):
     key = rsa.generate_private_key(public_exponent=65537, key_size=2048)
     pem = key.private_bytes(
         serialization.Encoding.PEM,
@@ -17,17 +16,16 @@ def test_issue_basic():
         serialization.NoEncryption(),
     )
     keyref = KeyRef(
-        kid="k1",
+        kid="k3",
         version=1,
         type=KeyType.RSA,
         uses=(KeyUse.SIGN,),
         export_policy=ExportPolicy.SECRET_WHEN_ALLOWED,
         material=pem,
     )
-    cert_bytes = SelfSignedCertificate().issue(keyref)
-    cert = x509.load_pem_x509_certificate(cert_bytes)
-    assert cert.subject == cert.issuer
-    assert (
-        cert.subject.get_attributes_for_oid(x509.oid.NameOID.COMMON_NAME)[0].value
-        == "localhost"
-    )
+    cert_builder = SelfSignedCertificate()
+
+    def issue():
+        cert_builder.issue(keyref)
+
+    benchmark(issue)
