+++ pkgs/standards/swarmauri_certs_self_signed/tests/unit/test_rfc5280_cert.py
@@ -1,14 +1,17 @@
+"""Tests for RFC 5280 compliance."""
+
 import pytest
 from cryptography.hazmat.primitives import serialization
 from cryptography.hazmat.primitives.asymmetric import rsa
+from cryptography import x509
 
-from swarmauri_standards import SelfSignedCertificate
+from swarmauri_certs_self_signed import SelfSignedCertificate
 from swarmauri_core.crypto.types import KeyRef, KeyType, KeyUse, ExportPolicy
 
 
 @pytest.mark.test
-@pytest.mark.perf
-def test_issue_perf(benchmark):
+@pytest.mark.unit
+def test_basic_constraints_and_key_usage():
     key = rsa.generate_private_key(public_exponent=65537, key_size=2048)
     pem = key.private_bytes(
         serialization.Encoding.PEM,
@@ -16,16 +19,19 @@ def test_issue_perf(benchmark):
         serialization.NoEncryption(),
     )
     keyref = KeyRef(
-        kid="k3",
+        kid="k4",
         version=1,
         type=KeyType.RSA,
         uses=(KeyUse.SIGN,),
         export_policy=ExportPolicy.SECRET_WHEN_ALLOWED,
         material=pem,
     )
-    cert_builder = SelfSignedCertificate()
-
-    def issue():
-        cert_builder.issue(keyref)
-
-    benchmark(issue)
+    cert_bytes = SelfSignedCertificate.tls_server(
+        "example.com", dns_names=["example.com"]
+    ).issue(keyref)
+    cert = x509.load_pem_x509_certificate(cert_bytes)
+    bc = cert.extensions.get_extension_for_class(x509.BasicConstraints)
+    assert bc.value.ca is False
+    ku = cert.extensions.get_extension_for_class(x509.KeyUsage)
+    assert ku.value.digital_signature
+    assert not ku.value.crl_sign
