--- pkgs/standards/swarmauri_standards/swarmauri_standards/__init__.py
@@ -1,3 +1,0 @@
-from .certs.SelfSignedCertificate import SelfSignedCertificate
-
-__all__ = ["SelfSignedCertificate"]
